+++ widgets/button/button_test.go
@@ -65,16 +65,22 @@ func (ct *callbackTracker) callback() error {
 
 func TestButton(t *testing.T) {
 	tests := []struct {
-		desc         string
-		text         string
-		callback     *callbackTracker
-		opts         []Option
-		events       []terminalapi.Event
-		canvas       image.Rectangle
-		want         func(size image.Point) *faketerm.Terminal
-		wantCallback *callbackTracker
-		wantNewErr   bool
-		wantDrawErr  bool
+		desc     string
+		text     string
+		callback *callbackTracker
+		opts     []Option
+		events   []terminalapi.Event
+		canvas   image.Rectangle
+
+		// timeSince is used to replace time.Since for tests, leave nil to use
+		// the original.
+		timeSince func(time.Time) time.Duration
+
+		want            func(size image.Point) *faketerm.Terminal
+		wantCallback    *callbackTracker
+		wantNewErr      bool
+		wantDrawErr     bool
+		wantCallbackErr bool
 	}{
 		{
 			desc:       "New fails with nil callback",
@@ -107,6 +113,13 @@ func TestButton(t *testing.T) {
 			},
 			canvas:     image.Rect(0, 0, 1, 1),
 			wantNewErr: true,
+		},
+		{
+			desc:        "draw fails on canvas too small",
+			callback:    &callbackTracker{},
+			text:        "hello",
+			canvas:      image.Rect(0, 0, 1, 1),
+			wantDrawErr: true,
 		},
 		{
 			desc:     "draws button in up state",
@@ -259,23 +272,246 @@ func TestButton(t *testing.T) {
 			},
 			wantCallback: &callbackTracker{},
 		},
+		{
+			desc:     "keyboard event triggers the button, trigger time didn't expire so button is down",
+			callback: &callbackTracker{},
+			text:     "hello",
+			opts: []Option{
+				Key(keyboard.KeyEnter),
+			},
+			timeSince: func(time.Time) time.Duration {
+				return 200 * time.Millisecond
+			},
+			canvas: image.Rect(0, 0, 8, 4),
+			events: []terminalapi.Event{
+				&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				cvs := testcanvas.MustNew(ft.Area())
 
-		// Keyboard event ignored when no key configured
-		// Draws button down by key + trigger.
-		// Releases button after key press.
-		// Doesn't release button after key press if before KeyUpDelay.
-		// Ignores unrelated key.
-		// Key works when KeyScopeFocused.
-		// sets custom key
-		// Ignores key outside of the container on KeyScopeFocused.
-		// Accepts key outside of the container on KeyScopeFlobal.
-		// Triggers callback multiple times.
-		// Callback returns an error.
-		// Custom height.
-		// Different width due to text.
-		// Different width due to WidthFor.
-		// Trims text on custom width.
+				// Button.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(1, 1, 8, 4), 'x', cell.BgColor(cell.ColorNumber(117)))
 
+				// Text.
+				testdraw.MustText(cvs, "hello", image.Point{2, 2},
+					draw.TextCellOpts(
+						cell.FgColor(cell.ColorBlack),
+						cell.BgColor(cell.ColorNumber(117))),
+				)
+
+				testcanvas.MustApply(cvs, ft)
+				return ft
+			},
+			wantCallback: &callbackTracker{
+				called: true,
+				count:  1,
+			},
+		},
+		{
+			desc:     "keyboard event triggers the button, custom trigger time expired so button is up",
+			callback: &callbackTracker{},
+			text:     "hello",
+			opts: []Option{
+				Key(keyboard.KeyEnter),
+				KeyUpDelay(100 * time.Millisecond),
+			},
+			timeSince: func(time.Time) time.Duration {
+				return 200 * time.Millisecond
+			},
+			canvas: image.Rect(0, 0, 8, 4),
+			events: []terminalapi.Event{
+				&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				cvs := testcanvas.MustNew(ft.Area())
+
+				// Shadow.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(1, 1, 8, 4), 's', cell.BgColor(cell.ColorNumber(240)))
+
+				// Button.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(0, 0, 7, 3), 'x', cell.BgColor(cell.ColorNumber(117)))
+
+				// Text.
+				testdraw.MustText(cvs, "hello", image.Point{1, 1},
+					draw.TextCellOpts(
+						cell.FgColor(cell.ColorBlack),
+						cell.BgColor(cell.ColorNumber(117))),
+				)
+
+				testcanvas.MustApply(cvs, ft)
+				return ft
+			},
+			wantCallback: &callbackTracker{
+				called: true,
+				count:  1,
+			},
+		},
+		{
+			desc:     "keyboard event triggers the button multiple times",
+			callback: &callbackTracker{},
+			text:     "hello",
+			opts: []Option{
+				Key(keyboard.KeyEnter),
+				KeyUpDelay(100 * time.Millisecond),
+			},
+			timeSince: func(time.Time) time.Duration {
+				return 200 * time.Millisecond
+			},
+			canvas: image.Rect(0, 0, 8, 4),
+			events: []terminalapi.Event{
+				&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+				&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+				&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				cvs := testcanvas.MustNew(ft.Area())
+
+				// Shadow.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(1, 1, 8, 4), 's', cell.BgColor(cell.ColorNumber(240)))
+
+				// Button.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(0, 0, 7, 3), 'x', cell.BgColor(cell.ColorNumber(117)))
+
+				// Text.
+				testdraw.MustText(cvs, "hello", image.Point{1, 1},
+					draw.TextCellOpts(
+						cell.FgColor(cell.ColorBlack),
+						cell.BgColor(cell.ColorNumber(117))),
+				)
+
+				testcanvas.MustApply(cvs, ft)
+				return ft
+			},
+			wantCallback: &callbackTracker{
+				called: true,
+				count:  3,
+			},
+		},
+		{
+			desc:     "mouse event triggers the button multiple times",
+			callback: &callbackTracker{},
+			text:     "hello",
+			canvas:   image.Rect(0, 0, 8, 4),
+			events: []terminalapi.Event{
+				&terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonLeft},
+				&terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonRelease},
+				&terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonLeft},
+				&terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonRelease},
+			},
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				cvs := testcanvas.MustNew(ft.Area())
+
+				// Shadow.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(1, 1, 8, 4), 's', cell.BgColor(cell.ColorNumber(240)))
+
+				// Button.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(0, 0, 7, 3), 'x', cell.BgColor(cell.ColorNumber(117)))
+
+				// Text.
+				testdraw.MustText(cvs, "hello", image.Point{1, 1},
+					draw.TextCellOpts(
+						cell.FgColor(cell.ColorBlack),
+						cell.BgColor(cell.ColorNumber(117))),
+				)
+
+				testcanvas.MustApply(cvs, ft)
+				return ft
+			},
+			wantCallback: &callbackTracker{
+				called: true,
+				count:  2,
+			},
+		},
+		{
+			desc: "the callback returns an error after a mouse event",
+			callback: &callbackTracker{
+				wantErr: true,
+			},
+			text:   "hello",
+			canvas: image.Rect(0, 0, 8, 4),
+			events: []terminalapi.Event{
+				&terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonLeft},
+				&terminalapi.Mouse{Position: image.Point{0, 0}, Button: mouse.ButtonRelease},
+			},
+			wantCallbackErr: true,
+		},
+		{
+			desc: "the callback returns an error after a keyboard event",
+			callback: &callbackTracker{
+				wantErr: true,
+			},
+			text: "hello",
+			opts: []Option{
+				Key(keyboard.KeyEnter),
+				KeyUpDelay(100 * time.Millisecond),
+			},
+			timeSince: func(time.Time) time.Duration {
+				return 200 * time.Millisecond
+			},
+			canvas: image.Rect(0, 0, 8, 4),
+			events: []terminalapi.Event{
+				&terminalapi.Keyboard{Key: keyboard.KeyEnter},
+			},
+			wantCallbackErr: true,
+		},
+		{
+			desc:     "draws button with custom height (infra gives smaller canvas)",
+			callback: &callbackTracker{},
+			text:     "hello",
+			canvas:   image.Rect(0, 0, 8, 2),
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				cvs := testcanvas.MustNew(ft.Area())
+
+				// Shadow.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(1, 1, 8, 2), 's', cell.BgColor(cell.ColorNumber(240)))
+
+				// Button.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(0, 0, 7, 1), 'x', cell.BgColor(cell.ColorNumber(117)))
+
+				// Text.
+				testdraw.MustText(cvs, "hello", image.Point{1, 0},
+					draw.TextCellOpts(
+						cell.FgColor(cell.ColorBlack),
+						cell.BgColor(cell.ColorNumber(117))),
+				)
+
+				testcanvas.MustApply(cvs, ft)
+				return ft
+			},
+			wantCallback: &callbackTracker{},
+		},
+		{
+			desc:     "button width adjusts to width (infra gives smaller canvas)",
+			callback: &callbackTracker{},
+			text:     "h",
+			canvas:   image.Rect(0, 0, 4, 2),
+			want: func(size image.Point) *faketerm.Terminal {
+				ft := faketerm.MustNew(size)
+				cvs := testcanvas.MustNew(ft.Area())
+
+				// Shadow.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(1, 1, 4, 2), 's', cell.BgColor(cell.ColorNumber(240)))
+
+				// Button.
+				testcanvas.MustSetAreaCells(cvs, image.Rect(0, 0, 3, 1), 'x', cell.BgColor(cell.ColorNumber(117)))
+
+				// Text.
+				testdraw.MustText(cvs, "h", image.Point{1, 0},
+					draw.TextCellOpts(
+						cell.FgColor(cell.ColorBlack),
+						cell.BgColor(cell.ColorNumber(117))),
+				)
+
+				testcanvas.MustApply(cvs, ft)
+				return ft
+			},
+			wantCallback: &callbackTracker{},
+		},
 		{
 			desc:     "sets custom text color",
 			callback: &callbackTracker{},
@@ -372,6 +608,12 @@ func TestButton(t *testing.T) {
 	shadowRune = 's'
 	for _, tc := range tests {
 		t.Run(tc.desc, func(t *testing.T) {
+			if tc.timeSince != nil {
+				timeSince = tc.timeSince
+			} else {
+				timeSince = time.Since
+			}
+
 			gotCallback := tc.callback
 			var cFn CallbackFn
 			if gotCallback == nil {
@@ -402,16 +644,38 @@ func TestButton(t *testing.T) {
 				}
 			}
 
-			for _, ev := range tc.events {
+			for i, ev := range tc.events {
 				switch e := ev.(type) {
 				case *terminalapi.Mouse:
-					if err := b.Mouse(e); err != nil {
-						t.Fatalf("Mouse => unexpected error: %v", err)
+					err := b.Mouse(e)
+					// Only the last event in test cases is the one that triggers the callback.
+					if i == len(tc.events)-1 {
+						if (err != nil) != tc.wantCallbackErr {
+							t.Errorf("Mouse => unexpected error: %v, wantCallbackErr: %v", err, tc.wantCallbackErr)
+						}
+						if err != nil {
+							return
+						}
+					} else {
+						if err != nil {
+							t.Fatalf("Mouse => unexpected error: %v", err)
+						}
 					}
 
 				case *terminalapi.Keyboard:
-					if err := b.Keyboard(e); err != nil {
-						t.Fatalf("Keyboard => unexpected error: %v", err)
+					err := b.Keyboard(e)
+					// Only the last event in test cases is the one that triggers the callback.
+					if i == len(tc.events)-1 {
+						if (err != nil) != tc.wantCallbackErr {
+							t.Errorf("Keyboard => unexpected error: %v, wantCallbackErr: %v", err, tc.wantCallbackErr)
+						}
+						if err != nil {
+							return
+						}
+					} else {
+						if err != nil {
+							t.Fatalf("Keyboard => unexpected error: %v", err)
+						}
 					}
 
 				default: